+++ src/Components/Home/Home.js
@@ -12,12 +12,18 @@ import Body from '../Body/Body';
 import Mind from '../Mind/Mind';
 import Productivity from '../Productivity/Productivity';
 import './Home.css';
+import useTodos from '../../hooks/useTodos';
 
 export default function Home() {
   const { user } = useContext(UserContext);
+  const { loading } = useTodos();
 
   if (!user) {
     return <Redirect to="/auth/sign-in" />;
+  }
+
+  if (loading) {
+    return <h1>Loading</h1>;
   }
 
   return (
